rename cache objects to cut / slice objects
